requirements.txt
Flask
Flask_Cors
mysql-connector-python
gunicorn